chore: reorganize Models into single folder
