+++ LotCoMPrinter/Models/Datasources/JBKQueue.cs
@@ -1,4 +1,4 @@
-namespace LotCoMPrinter.Models.Serialization;
+namespace LotCoMPrinter.Models.Datasources;
 
 public partial class JBKQueue() : 
     SerialQueue(QueuePath: "\\\\144.133.122.1\\Lot Control Management\\Database\\process_control\\serial_queues\\_jbk_queue.json",
